+++ package_tests/test_portlet.py
@@ -1,6 +1,6 @@
 """Test view generation."""
 
-from .base import SETUPPY_TEMPLATE
+from .base import PYPROJECTTOML_TEMPLATE
 from bobtemplates.plone import base
 from bobtemplates.plone import portlet
 from mrbob.bobexceptions import ValidationError
@@ -496,8 +496,8 @@ class Renderer(base.Renderer):
     with open(os.path.join(package_path + "/configure.zcml"), "w") as f:
         f.write(template)
 
-    template = SETUPPY_TEMPLATE
-    with open(os.path.join(target_path + "/setup.py"), "w") as f:
+    template = PYPROJECTTOML_TEMPLATE
+    with open(os.path.join(target_path + "/pyproject.toml"), "w") as f:
         f.write(template)
 
     template = """<?xml version="1.0"?>