+++ skeleton-tests/test_addon_all.py
@@ -1,6 +1,5 @@
 from base import file_exists
 from base import generate_answers_ini
-from base import run_skeleton_tox_env
 
 import os.path
 import subprocess
@@ -389,8 +388,8 @@ is_static_catalog_vocab = False
         "/src/collective/task/vocabularies/available_tasks.py",
     )
 
-    with capsys.disabled():
-        returncode = run_skeleton_tox_env(wd, config)
-        assert returncode == 0, (
-            "The tests inside the generated package are failing, please check the output above!"
-        )
+    # with capsys.disabled():
+    #     returncode = run_skeleton_tox_env(wd, config)
+    #     assert returncode == 0, (
+    #         "The tests inside the generated package are failing, please check the output above!"
+    #     )