+++ package_tests/test_viewlet.py
@@ -1,6 +1,6 @@
 """Test view generation."""
 
-from .base import SETUPPY_TEMPLATE
+from .base import PYPROJECTTOML_TEMPLATE
 from bobtemplates.plone import base
 from bobtemplates.plone import viewlet
 from mrbob.bobexceptions import SkipQuestion
@@ -359,8 +359,8 @@ class MyViewlet(ViewletBase):
         },
     )
 
-    template = SETUPPY_TEMPLATE
-    with open(os.path.join(target_path + "/setup.py"), "w") as f:
+    template = PYPROJECTTOML_TEMPLATE
+    with open(os.path.join(target_path + "/pyproject.toml"), "w") as f:
         f.write(template)
 
     os.chdir(package_path)
@@ -456,8 +456,8 @@ class MyViewlet(ViewletBase):
         },
     )
 
-    template = SETUPPY_TEMPLATE
-    with open(os.path.join(target_path + "/setup.py"), "w") as f:
+    template = PYPROJECTTOML_TEMPLATE
+    with open(os.path.join(target_path + "/pyproject.toml"), "w") as f:
         f.write(template)
 
     os.chdir(package_path)