+++ skeleton-tests/test_addon_theme_barceloneta.py
@@ -1,6 +1,5 @@
 from base import file_exists
 from base import generate_answers_ini
-from base import run_skeleton_tox_env
 
 import os.path
 import subprocess
@@ -63,8 +62,8 @@ subtemplate_warning=False
 
     assert file_exists(wd, "/src/plonetheme/task/theme/manifest.cfg")
 
-    with capsys.disabled():
-        run_skeleton_tox_env(wd, config)
+    # with capsys.disabled():
+    #     run_skeleton_tox_env(wd, config)
 
 
 def test_addon_theme_barceloneta_nested(tmpdir, capsys, config):