+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 7.0.0a4 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- editorconfig: Don't use inline comments.
+  Inline comments are invalid since editorconfig v0.15.0.
+  [thet]
 
 
 7.0.0a3 (2025-07-22)