+++ CHANGES.rst
@@ -4,8 +4,15 @@ Changelog
 7.0.0a4 (unreleased)
 --------------------
 
+- Refactor all tests to adapt to the new package structure created by cookieplone
+  [erral]
+
+- Use cookieplone to create an addon
+  [erral]
+
 - Allow overriding browser layer in bobtemplates.cfg
   [erral]
+
 
 7.0.0a3 (2025-07-22)
 --------------------